store: moved pkg core to go-textile-core
